Reorganize career profile edit page, fix input persistence
