+++ app/profile/career/edit/page.scss
@@ -1,4 +1,4 @@
-@use "../../styles/vars";
+@use "@/styles/vars";
 
 .career-profile-page {
   max-width: 1400px;
